Poppup clusters (#6)
For the data simulation we needed a problem that the clusters occurs over time. This is relevant and synonymous to the fact that fraude also changes over time and new types of fraud occur. With the popping_clusters module one is able to simulate this in the most basic 2d form.
In the popping_clusters.py file are the functionalities described to generate a data set with popping clusters.

One call a example dataset by simply importing popping_clusters and running cluster_generator. As output one gets an array of dicts with the information on x, y, time and cluster label.

The clusters are generated via random distributions. The clusters are random gaussians in X and Y which are not correlated. The probability of a cluster popping-up will change over time. This can be generated via 2 functions, moving sinuses or a moving poisson.