Poppup clusters (#9)
Created a popping_clusters module and converted into a generator. With this module it is possible to generate popping-up clusters in the xy plane. Also an demo notebook with documentation is created to show the working of this module.

Also some restructering of documents with respect to the popping clusters is done.